customer-service/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.6
Django==5.2.12
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
mysqlclient==2.2.8
PyJWT==2.12.1
python-dotenv==1.2.2
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
